requirements.txt
"""
Requirements.txt - All Python dependencies for the backend
"""

Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.0
gunicorn==21.0.0
